web/requirements.txt
asgiref==3.2.3
attrs==19.3.0
Babel==2.8.0
click==7.1.1
click-plugins==1.1.1
cligj==0.5.0
Django==2.2
django-address==0.2.1
django-extensions==2.2.8
django-mysql==3.3.0
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.0.3
Fiona==1.8.13.post1
geographiclib==1.50
geopandas==0.7.0
geopy==1.21.0
gunicorn==20.0.4
importlib-metadata==1.6.0
more-itertools==8.2.0
munch==2.5.0
mysqlclient==1.4.6
numpy==1.18.1
packaging==20.3
pandas==1.0.1
phonenumbers==8.11.2
pluggy==0.13.1
py==1.8.1
pycountry==19.8.18
pyparsing==2.4.7
pyproj==2.5.0
pytest==5.4.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.2.0
Shapely==1.7.0
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
wcwidth==0.1.9
zipp==3.1.0
